Minor bug fixes (described below)
- cmd_line.cpp
  Fix segmentation fault if user enters interface number 1 greater than max or "0". (Example: program printed "Enter the interface number (1-2):" but user entered "3", this would segfault)

- cli_command_format.cpp
  Fix compiler warning (sign-compare)

- cmd_line_main.cpp
  Fix compiler warning: Unused variable "show_cmd_separator"

+++ controller/app/cmdline/src/cli_command_format.cpp
@@ -70,7 +70,7 @@ bool cli_command_format::run_command(cmd_line * cmd_ptr, std::queue<std::string>
             match_count++;
             total_matched++;
 
-            if ((match_max != UINT_MAX) && (match_count >= match_max))
+            if ((match_max != static_cast<int>(UINT_MAX)) && (match_count >= match_max))
             {
                 // Move to next argument - completed this one
                 m_i++;
